@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
-{
-    public class AlbumTracks
-    {
-        public List<Track> items { get; set; }
+namespace SpotifyDemo1.Objects;
 
-    }
+public class AlbumTracks
+{
+    public List<Track> items { get; set; }
+
 }
