@@ -2,15 +2,14 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
-{
-    public class ArtistSearch
-    {
-        public Artists artists { get; set; }
+namespace SpotifyDemo1.Objects;
 
-        public class Artists
-        {
-            public List<Artist> items { get; set; }
-        }
+public class ArtistSearch
+{
+    public Artists artists { get; set; }
+
+    public class Artists
+    {
+        public List<Artist> items { get; set; }
     }
 }
