@@ -2,10 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
+namespace SpotifyDemo1.Objects;
+
+public class PlaylistMeta
 {
-    public class PlaylistMeta
-    {
-        public List<PlaylistTrackMeta> items { get; set; }
-    }
+    public List<PlaylistTrackMeta> items { get; set; }
 }
