@@ -2,16 +2,15 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
+namespace SpotifyDemo1.Objects;
+
+public class Token
 {
-    public class Token
-    {
-        public string access_token { get; set; }
+    public string access_token { get; set; }
 
-        public string token_type { get; set; }
+    public string token_type { get; set; }
 
-        public int expires_in { get; set; }
+    public int expires_in { get; set; }
 
-        public string scope { get; set; }
-    }
+    public string scope { get; set; }
 }
