@@ -2,24 +2,23 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
+namespace SpotifyDemo1.Objects;
+
+public class Track : IEquatable<Track>
 {
-    public class Track : IEquatable<Track>
-    {
-        public string id { get; set; }
+    public string id { get; set; }
 
-        public string name { get; set; }
+    public string name { get; set; }
 
-        public int? popularity { get; set; }
+    public int? popularity { get; set; }
 
-        public Album album { get; set; }
+    public Album album { get; set; }
 
-        public AudioFeatures audioFeatures { get; set; }
+    public AudioFeatures audioFeatures { get; set; }
 
-        public bool Equals(Track other) =>
-            (id ?? "0").Equals(other.id ?? "0");
+    public bool Equals(Track other) =>
+        (id ?? "0").Equals(other.id ?? "0");
 
-        public override int GetHashCode() =>
-            (id ?? "0").GetHashCode();
-    }
+    public override int GetHashCode() =>
+        (id ?? "0").GetHashCode();
 }
