@@ -2,11 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SpotifyDemo1.Objects
-{
-    public class TrackMeta
-    {
-        public int total { get; set; }
-    }
+namespace SpotifyDemo1.Objects;
 
+public class TrackMeta
+{
+    public int total { get; set; }
 }
